+++ packages/ethereum-indexer/src/internal/types.ts
@@ -1,5 +1,3 @@
-import {Abi} from 'abitype';
-import {LogEvent} from '../types';
 
 export type JSONObject = {
 	[key: string]: JSONType;